Added manager script to easily drop or create the Teamwork database, adjusted sessionmaker so that we're only creating it once and invoking Session() everywhere else as recommended by the docs, and modified requirements.txt to reflect the new changes

+++ jobs/models.py
@@ -38,12 +38,13 @@ def connect_to_database():
     return create_engine(settings.DB_STRING, echo=settings.DEBUG)
 
 
-def create_database(engine):
+def create_database(engine, auto_drop=False):
     """
     Setup the database (for deployment)
     """
-    Base.metadata.drop_all(bind=engine, checkfirst=True)
-    Base.metadata.create_all(engine)
+    if auto_drop is True:
+        Base.metadata.drop_all(bind=engine, checkfirst=True)
+    Base.metadata.create_all(engine, checkfirst=True)
 
 
 class TWProject(Base):
